+++ root_orchestrator/resource-abstractor/resources_blueprint.py
